refactor: reconstructing the components
