+++ backend/app/admin_routes.py
@@ -80,124 +80,6 @@ def logout():
     return redirect(url_for('home'))
 
 
-@admin_bp.route('/partners', methods=['GET'])
-def get_partners():
-    """fetches all the parteners in the DB"""
-    try:
-        partners = Patners.query.all()
-        partners_data = [{"id": partner.id, "name": partner.name,
-                          "image": partner.image} for partner in partners]
-        return jsonify(partners_data), 200
-    except Exception as e:
-        return jsonify({"error": str(e)}), 400
-
-
-@admin_bp.route('/partner/new/', methods=['POST'])
-@jwt_required()
-def add_partner():
-    """adding partners in the footer secion"""
-    try:
-        data = request.json
-
-        partner_data = {
-            'name': data.get('name'),
-            'image': data.get('image_base64')
-        }
-
-        new_partner = Patners(**partner_data)
-        db.session.add(new_partner)
-        db.session.commit()
-
-        return jsonify(
-            {
-                "status": "success",
-                "message": "Partner added successfully"
-            }
-        ), 201
-
-    except Exception as e:
-        db.session.rollback()
-        return jsonify({'error': str(e)}), 500
-
-
-@admin_bp.route('/partner/delete/<int:partner_id>/', methods=['DELETE'])
-@jwt_required()
-def delete_partner(partner_id):
-    """deletes a partner from the DB
-
-    Args:
-        partner_id (int): the ID of the partner
-
-    """
-    try:
-        partner = Patners.query.get(partner_id)
-        if partner:
-            db.session.delete(partner)
-            db.session.commit()
-            return '', 204
-        return jsonify({"message": "Partner not found!"}), 404
-    except Exception as e:
-        db.session.rollback()
-        return jsonify({"error": str(e)}), 400
-
-
-@admin_bp.route('/social/', methods=['GET'])
-def get_social():
-    """gets all the social media accounts"""
-    social = Social.query.all()
-    social_list = []
-    for soc in social:
-        social_list.append({
-            'twitter': soc.twitter,
-            'email': soc.email,
-            'instagram': soc.instagram,
-            'phone': soc.phone,
-
-        })
-    return jsonify({"social": social_list})
-
-
-@admin_bp.route('/social/new/', methods=['POST'])
-@jwt_required()
-def add_social():
-    """creates a new social media accounts
-
-    Returns:
-        json: returns json data for the social media accounts
-    """
-    try:
-        data = request.get_json()
-        new_social = Social(
-            phone=data['phone'],
-            twitter=data['twitter'],
-            instagram=data['instagram'],
-            email=data['email']
-        )
-        db.session.add(new_social)
-        db.session.commit()
-        return jsonify({"message": "social added successfully!"}), 201
-    except Exception as e:
-        db.session.rollback()
-        return jsonify({"error": str(e)}), 500
-
-
-@admin_bp.route('/social/delete/', methods=['DELETE'])
-@jwt_required()
-def delete_social():
-    """deletes all social media account"""
-    try:
-        social_records = Social.query.all()
-
-        if social_records:
-            for social in social_records:
-                db.session.delete(social)
-            db.session.commit()
-            return '', 204
-        return jsonify({"message": "No social records found"}), 404
-    except Exception as e:
-        return jsonify({"error": str(e)}), 500
-
-
 @admin_bp.route('/credentials/<int:id>/', methods=['GET'])
 def get_credentials(id):
     """fetches the password and username"""
