+++ frontend/src/admin-components/DeleteSparePart.jsx
