--- frontend/src/components/CarLogos.jsx
@@ -1,88 +1,0 @@
-import React from "react";
-
-const CarLogos = () => {
-  return {
-    logo1: {
-      name: "BMW",
-      image: "../images/BMW-logo.png",
-    },
-    logo2: {
-      name: "Chevrolet",
-      image: "../images/chevrolet-logo.jpeg",
-    },
-    logo3: {
-      name: "Ford",
-      image: "../images/ford-logo.png",
-    },
-    logo4: {
-      name: "Honda",
-      image: "../images/honda-logo.jpeg",
-    },
-    logo5: {
-      name: "Isuzu",
-      image: "../images/isuzu-Logo.png",
-    },
-    logo6: {
-      name: "Jaguar",
-      image: "../images/jaguar-Logo.png",
-    },
-    logo7: {
-      name: "Jeep",
-      image: "../images/jeep-logo.jpeg",
-    },
-    logo8: {
-      name: "Land Rover",
-      image: "../images/landrover-logo.png",
-    },
-    logo9: {
-      name: "Mazda",
-      image: "../images/mazda-Logo.png",
-    },
-    logo10: {
-      name: "Mercedes",
-      image: "../images/mercedes-logo.webp",
-    },
-    logo11: {
-      name: "Mini",
-      image: "../images/mini-logo.png",
-    },
-    logo12: {
-      name: "Mitsubishi",
-      image: "../images/mitsubishi-logo.png",
-    },
-    logo13: {
-      name: "Nissan",
-      image: "../images/nissan-logo.jpeg",
-    },
-    logo14: {
-      name: "Peugeot",
-      image: "../images/peugot-logo.webp",
-    },
-    logo15: {
-      name: "Subaru",
-      image: "../images/subaru-logo.png",
-    },
-    logo16: {
-      name: "Suzuki",
-      image: "../images/suzuki-logo.png",
-    },
-    logo17: {
-      name: "Toyota",
-      image: "../images/toyota-logo.jpeg",
-    },
-    logo18: {
-      name: "Volvo",
-      image: "../images/volvo-logo.jpeg",
-    },
-    logo19: {
-      name: "VW",
-      image: "../images/vw-logo.png",
-    },
-    logo20: {
-      name: "Audi",
-      image: "../images/audi-logo.png",
-    },
-  };
-};
-
-export default CarLogos;
