--- frontend/src/components/ProtectedRoute.jsx
@@ -1,19 +1,0 @@
-import React from 'react';
-import { Route, Redirect } from 'react-router-dom';
-
-const PrivateRoute = ({ component: Component, isAuthenticated, ...rest }) => {
-    return (
-        <Route
-            {...rest}
-            render={(props) =>
-                isAuthenticated ? (
-                    <Component {...props} />
-                ) : (
-                    <Redirect to="/login" />
-                )
-            }
-        />
-    );
-};
-
-export default PrivateRoute;
